Created Player Animation Script
Move the players animations to another script to keep it simple. this is untested so no clue if it works or not well find out once we have all the animations in order. Player movement should all be working but that went all untested to also deleted the menu since I want to work on a single boss at a time and a menu would be consider polish IMO